+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fix #3600: add owner references support to HasMetadata
 
 #### Dependency Upgrade
+* Fix #3505: Update OpenShift Model to latest version (4.9.x)
 
 #### New Features
 * Fix #3579: Add support for Volcano extension